Refactor - move page/fragment classes to inner packages
