+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/pages/plr/facility/search/SearchFacilityIdFragment.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility;
+package ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.search;
 
 import ca.bc.gov.health.qa.autotest.plr.web.pages.components.DropDownMenu;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.SearchSectionFragment;
