+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/tests/ViewFacilitySimpleTests.java
@@ -1,19 +1,11 @@
 package ca.bc.gov.health.qa.autotest.plr.web.tests;
 
-import static ca.bc.gov.health.qa.autotest.plr.web.tests.TestHelper.navigateToSearchFacilityPage;
 import static java.util.Objects.requireNonNull;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.Logger;
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -21,17 +13,12 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import ca.bc.gov.health.qa.autotest.core.util.text.TextUtils;
 import ca.bc.gov.health.qa.autotest.plr.data.InjectableData;
 import ca.bc.gov.health.qa.autotest.plr.data.PlrData;
-import ca.bc.gov.health.qa.autotest.plr.util.ProviderType;
 import ca.bc.gov.health.qa.autotest.plr.util.UserType;
 import ca.bc.gov.health.qa.autotest.plr.web.actions.ViewFacilitySimpleActions;
-import ca.bc.gov.health.qa.autotest.plr.web.actions.model.facility.Identifier;
 import ca.bc.gov.health.qa.autotest.plr.web.actions.model.facility.Relationship;
-import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.ViewMode;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.FacilitySection;
-import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.SearchFacilityPage;
 import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.facility.ViewFacilityPage;
 //import ca.bc.gov.health.qa.autotest.plr.web.pages.plr.provider.ProviderSection;
 import ca.bc.gov.health.qa.autotest.plr.web.workflows.PlrWebWorkflow;
