+++ autotest-plr/src/main/java/ca/bc/gov/health/qa/autotest/plr/web/tests/types/ElectronicAddressType.java
@@ -1,0 +1,27 @@
+package ca.bc.gov.health.qa.autotest.plr.web.tests.types;
+
+public enum ElectronicAddressType {
+	EMAIL("E - Email"), 
+	FTP("F - FTP"),
+	HTTP("H - HTTP");
+
+	private String text;
+
+	ElectronicAddressType(String text) {
+		this.text = text;
+	}
+
+	public String getText() {
+		return this.text;
+	}
+
+	public static ElectronicAddressType fromString(String text) {
+		for (ElectronicAddressType b : ElectronicAddressType.values()) {
+			if (b.text.equalsIgnoreCase(text)) {
+				return b;
+			}
+		}
+		return null;
+	}
+
+}
